server/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.3.1
